move pledge stuff to pledgeable concern
