+++ app/models/concerns/urgency.rb
@@ -4,7 +4,7 @@ module Urgency
   def still_urgent?
     # Verify that a pregnancy has not been marked urgent in the past six days
     return false if recent_history_tracks.count == 0
-    return false if pregnancy.pledge_sent || pregnancy.resolved_without_dcaf
+    return false if pledge_sent || resolved_without_dcaf
     recent_history_tracks.sort.reverse.each do |history|
       return true if history.marked_urgent?
     end
@@ -31,5 +31,4 @@ module Urgency
   def recent_history_tracks
     history_tracks.select { |ht| ht.updated_at > 6.days.ago }
   end
-
 end
